Moved code from views to do
Also updated a couple send2middleware calls
