+++ monitor/do.py
@@ -6,6 +6,9 @@ from monitor.models import Archive
 from monitor.models import Config
 from monitor.models import Event
 
+from monitor.middleware import send2middleware
+
+from time import sleep
 import datetime
 import pytz
 
@@ -333,3 +336,42 @@ def getAllData(cur_beer):
     all_data = all_data + reading_data
 
     return all_data
+
+def get_date_diff(d1,d2, append = "ago"):
+
+    '''Returns the difference between two datetime objects in a readable format'''
+    diff = abs(d2-d1)
+
+    if(diff.days > 0): out = str(diff.days) + " day(s)"
+    elif(diff.seconds < 60): out = "less than a minute"
+    elif(diff.seconds < 60*60): out = str(int(round(diff.seconds/60,0))) + " minute(s)"
+    else: out = str(int(round(diff.seconds/(60*60),0))) + " hour(s)"
+
+    if append != None: out = out + " " + append
+    return(out)
+
+def get_paint_cols(val, rng = None):
+    '''Returns the background color for a value given a set range. In the future, it could also return foreground color'''
+    if rng == None or rng == (0,0): bgcol = "#FFFFFF" #White
+    elif(rng[0] <= val <= rng[1]): bgcol = "#008000" #Green
+    elif(not (rng[0] <= val <= rng[1])): bgcol = "#FF0000" #Red
+
+    fgcol = "#000000" #Black
+    return((bgcol, fgcol))
+
+def next_log_estimate():
+    last_reading = getLastReading(getActiveBeer()).instant_actual
+    log_freq = None
+    for i in range(10):
+        r, msg = send2middleware("?code=M")
+        print(r)
+        if r.upper() == "SUCCESS":
+            log_freq = int(msg.split("=")[1])
+            break
+        sleep(.1)
+    out = "unknown amount of time"
+    if log_freq != None:
+        next = last_reading + timedelta(minutes = log_freq)
+        now = nowInUtc()
+        if next >= now: out = get_date_diff(nowInUtc(), next, append = None)
+    return(out)
