+++ monitor/views.py
@@ -171,10 +171,10 @@ def dashboard(request):
             "pres_beer": cur_pres_beer
         },
         "bgcols" : {
-            "temp_amb": get_paint_cols(cur_temp_amb, temp_amb_rng)[0],
-            "temp_beer": get_paint_cols(cur_temp_beer, temp_beer_rng)[0],
-            "light_amb": get_paint_cols(cur_light_amb)[0],
-            "pres_beer": get_paint_cols(cur_pres_beer)[0]
+            "temp_amb": do.get_paint_cols(cur_temp_amb, temp_amb_rng)[0],
+            "temp_beer": do.get_paint_cols(cur_temp_beer, temp_beer_rng)[0],
+            "light_amb": do.get_paint_cols(cur_light_amb)[0],
+            "pres_beer": do.get_paint_cols(cur_pres_beer)[0]
         },
         "greenrng": {
             "temp_amb": temp_amb_rng,
@@ -182,36 +182,13 @@ def dashboard(request):
         },
         "last_log_date": cur_reading.instant_actual.strftime("%Y-%m-%d"),
         "last_log_time": cur_reading.instant_actual.strftime("%H:%M:%S"),
-        "last_log_ago": get_date_diff(cur_reading.instant_actual, do.nowInUtc()),
-        "next_log": next_log_estimate(),
+        "last_log_ago": do.get_date_diff(cur_reading.instant_actual, do.nowInUtc()),
+        "next_log": do.next_log_estimate(),
         "all_beers": Beer.objects.all(),
         "active_beer": active_beer,
         "beer_date": active_beer.brew_date,
     }
     return render_to_response('dashboard.html',data)
-
-def get_date_diff(d1,d2, append = "ago"):
-
-    '''Returns the difference between two datetime objects in a readable format'''
-    diff = abs(d2-d1)
-
-    if(diff.days > 0): out = str(diff.days) + " day(s)"
-    elif(diff.seconds < 60): out = "less than a minute"
-    elif(diff.seconds < 60*60): out = str(int(round(diff.seconds/60,0))) + " minute(s)"
-    else: out = str(int(round(diff.seconds/(60*60),0))) + " hour(s)"
-
-    if append != None: out = out + " " + append
-    return(out)
-
-def get_paint_cols(val, rng = None):
-    '''Returns the background color for a value given a set range. In the future, it could also return foreground color'''
-    if rng == None or rng == (0,0): bgcol = "#FFFFFF" #White
-    elif(rng[0] <= val <= rng[1]): bgcol = "#008000" #Green
-    elif(not (rng[0] <= val <= rng[1])): bgcol = "#FF0000" #Red
-
-
-    fgcol = "#000000" #Black
-    return((bgcol, fgcol))
 
 def dashboard_update(request):
     '''Forces a log then returns to dashboard page'''
@@ -219,7 +196,7 @@ def dashboard_update(request):
     old_reading = do.getLastReading(active_beer)
 
     for i in range(4):
-        command_status = send2middleware("F")
+        command_status = send2middleware("?code=F")
         if command_status[0] == "Success": break
         sleep(.1)
     if command_status[0] == "Success":
@@ -227,17 +204,6 @@ def dashboard_update(request):
             if do.getLastReading(active_beer) != old_reading: break
             sleep(.1)
     return HttpResponseRedirect(reverse('dashboard'))
-
-def gen_unableToLoad(page_name, cur_beer):
-    '''Creates a page saying the page_name for cur_berr was unable to be loaded due to no readings'''
-    #In future: allow the message (reason for unable to load page) a parameter
-    data = {
-        'all_beers': Beer.objects.all(),
-        'active_beer': do.getActiveBeer(),
-        'page_name': page_name,
-        'cur_beer': cur_beer,
-    }
-    return render_to_response('unabletoload.html',data)
 
 def chart(request, cur_beer = None):
     '''Creates a chart page with the Google Annotation Chart'''
@@ -298,19 +264,14 @@ def data_chk(request, page_name = "dashboard", cur_beer = None):
         else: out = "404 Page Not Found"
 
     return out
-def next_log_estimate():
-    last_reading = do.getLastReading(do.getActiveBeer()).instant_actual
-    log_freq = None
-    for i in range(10):
-        r, msg = send2middleware("M")
-        print(r)
-        if r.upper() == "SUCCESS":
-            log_freq = int(msg.split("=")[1])
-            break
-        sleep(.1)
-    out = "unknown amount of time"
-    if log_freq != None:
-        next = last_reading + timedelta(minutes = log_freq)
-        now = do.nowInUtc()
-        if next >= now: out = get_date_diff(do.nowInUtc(), next, append = None)
-    return(out)
+
+def gen_unableToLoad(page_name, cur_beer):
+    '''Creates a page saying the page_name for cur_berr was unable to be loaded due to no readings'''
+    #In future: allow the message (reason for unable to load page) a parameter
+    data = {
+        'all_beers': Beer.objects.all(),
+        'active_beer': do.getActiveBeer(),
+        'page_name': page_name,
+        'cur_beer': cur_beer,
+    }
+    return render_to_response('unabletoload.html',data)
